services/contract-service/requirements.txt
# Contract Service Dependencies
Flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.1
PyJWT==2.8.0
requests==2.31.0
python-consul==1.1.0
